% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r
\name{add}
\alias{add}
\title{Add together two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{a numeric value}

\item{y}{a nuemric value}
}
\value{
the sum of the two provided  parameters
}
\description{
Add together two numbers
}
\examples{
add(10,13)
add(12,-34)

}
